agent-service/requirements.txt
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0
jinja2>=3.1.0
langchain>=0.3.0
langchain-ollama>=0.1.0
langchain-core>=0.3.0
playwright>=1.50.0